3.0.0
I need to stay on top of updating this.  Character selection is
finalized.  Project preview is working.  I have a new navbar, of which I
have it changing pages every so often.  The "About Me" page doesn't
change, but everything else will, and I have a function to revert back
to the original screen when you push "home", and it's working
beautifully.

What isn't working so well is the dark effect when you preview a
project.  I also need to specify that you can click on the image to
preview the project.